+++ CHANGELOG.md
@@ -5,6 +5,7 @@ UNRELEASED
 * [ [#1103](https://github.com/digitalfabrik/integreat-cms/issues/1103) ] Add automatic translations via DeepL API
 * [ [#1024](https://github.com/digitalfabrik/integreat-cms/issues/1024) ] Add URL search-replace for linkchecker
 * [ [#1177](https://github.com/digitalfabrik/integreat-cms/issues/1177) ] Add content locking mechanism
+* [ [#1255](https://github.com/digitalfabrik/integreat-cms/issues/1255) ] Check only the latest versions of translations for broken links
 
 
 2022.3.6