+++ CHANGELOG.md
@@ -6,6 +6,7 @@ UNRELEASED
 * [ [#1024](https://github.com/digitalfabrik/integreat-cms/issues/1024) ] Add URL search-replace for linkchecker
 * [ [#1177](https://github.com/digitalfabrik/integreat-cms/issues/1177) ] Add content locking mechanism
 * [ [#1255](https://github.com/digitalfabrik/integreat-cms/issues/1255) ] Check only the latest versions of translations for broken links
+* [ [#1054](https://github.com/digitalfabrik/integreat-cms/issues/1054) ] Provide fallback translations for mirrored pages
 
 
 2022.3.6