+++ CHANGELOG.md
@@ -2,12 +2,13 @@ UNRELEASED
 ----------
 
 * [ [#1319](https://github.com/digitalfabrik/integreat-cms/issues/1319) ] Fix error on Imprint API
-* [ [#1103](https://github.com/digitalfabrik/integreat-cms/issues/1103) ] Add automatic translations via DeepL API
+* [ [#1104](https://github.com/digitalfabrik/integreat-cms/issues/1104) ] Add automatic translations via DeepL API
 * [ [#1024](https://github.com/digitalfabrik/integreat-cms/issues/1024) ] Add URL search-replace for linkchecker
 * [ [#1177](https://github.com/digitalfabrik/integreat-cms/issues/1177) ] Add content locking mechanism
 * [ [#1255](https://github.com/digitalfabrik/integreat-cms/issues/1255) ] Check only the latest versions of translations for broken links
 * [ [#1054](https://github.com/digitalfabrik/integreat-cms/issues/1054) ] Provide fallback translations for mirrored pages
-* [ [#1198](https://github.com/digitalfabrik/integreat-cms/issues/1198) ] Check availability for DeepL bulk actions 
+* [ [#1198](https://github.com/digitalfabrik/integreat-cms/issues/1198) ] Check availability for DeepL bulk actions
+* [ [#1293](https://github.com/digitalfabrik/integreat-cms/issues/1293) ] Enable login via email address
 
 
 2022.3.6